+++ applications/app_balance.c
@@ -32,6 +32,7 @@
 #include "datatypes.h"
 #include "comm_can.h"
 #include "terminal.h"
+#include "digital_filter.h"
 
 
 #include <math.h>
@@ -68,16 +69,6 @@ typedef enum {
 	HALF,
 	ON
 } SwitchState;
-
-typedef struct{
-	float a0, a1, a2, b1, b2;
-	float z1, z2;
-} Biquad;
-
-typedef enum {
-	BQ_LOWPASS,
-	BQ_HIGHPASS
-} BiquadType;
 
 // Balance thread
 static THD_FUNCTION(balance_thread, arg);
@@ -136,35 +127,6 @@ static void terminal_experiment(int argc, const char **argv);
 static float app_balance_get_debug(int index);
 static void app_balance_sample_debug(void);
 static void app_balance_experiment(void);
-
-// Utility Functions
-float biquad_process(Biquad *biquad, float in) {
-    float out = in * biquad->a0 + biquad->z1;
-    biquad->z1 = in * biquad->a1 + biquad->z2 - biquad->b1 * out;
-    biquad->z2 = in * biquad->a2 - biquad->b2 * out;
-    return out;
-}
-void biquad_config(Biquad *biquad, BiquadType type, float Fc) {
-	float K = tanf(M_PI * Fc);	// -0.0159;
-	float Q = 0.707; // maximum sharpness (0.5 = maximum smoothness)
-	float norm = 1 / (1 + K / Q + K * K);
-	if (type == BQ_LOWPASS) {
-		biquad->a0 = K * K * norm;
-		biquad->a1 = 2 * biquad->a0;
-		biquad->a2 = biquad->a0;
-	}
-	else if (type == BQ_HIGHPASS) {
-		biquad->a0 = 1 * norm;
-		biquad->a1 = -2 * biquad->a0;
-		biquad->a2 = biquad->a0;
-	}
-	biquad->b1 = 2 * (K * K - 1) * norm;
-	biquad->b2 = (1 - K / Q + K * K) * norm;
-}
-void biquad_reset(Biquad *biquad) {
-	biquad->z1 = 0;
-	biquad->z2 = 0;
-}
 
 // Exposed Functions
 void app_balance_configure(balance_config *conf, imu_config *conf2) {
