+++ applications/appconf_default.h
@@ -554,6 +554,9 @@
 #ifndef APPCONF_IMU_FILTER
 #define APPCONF_IMU_FILTER					IMU_FILTER_LOW
 #endif
+#ifndef APPCONF_IMU_ACCEL_LOWPASS_FILTER
+#define APPCONF_IMU_ACCEL_LOWPASS_FILTER	0.0
+#endif
 #ifndef APPCONF_IMU_SAMPLE_RATE_HZ
 #define APPCONF_IMU_SAMPLE_RATE_HZ			200
 #endif
