+++ util/digital_filter.h
@@ -22,6 +22,17 @@
 
 #include <stdint.h>
 
+
+typedef struct{
+	float a0, a1, a2, b1, b2;
+	float z1, z2;
+} Biquad;
+
+typedef enum {
+	BQ_LOWPASS,
+	BQ_HIGHPASS
+} BiquadType;
+
 // Functions
 void filter_fft(int dir, int m, float *real, float *imag);
 void filter_dft(int dir, int len, float *real, float *imag);
@@ -31,5 +42,8 @@ void filter_zeroPad(float *data, float *result, int dataLen, int resultLen);
 void filter_create_fir_lowpass(float *filter_vector, float f_break, int bits, int use_hamming);
 float filter_run_fir_iteration(float *vector, float *filter, int bits, uint32_t offset);
 void filter_add_sample(float *buffer, float sample, int bits, uint32_t *offset);
+float biquad_process(Biquad *biquad, float in);
+void biquad_config(Biquad *biquad, BiquadType type, float Fc);
+void biquad_reset(Biquad *biquad);
 
 #endif /* DIGITAL_FILTER_H_ */
